Alpha - 0.03 - Started to add in-gameUI
With some new In-Round UI and added the ability to use stamina. The more the player attacks the more they loose stamina with a 5sec penalty if they use it all up. To buffer this, if the player stands still, they regain x2 stamina back. In addition, the Health UI works.