Alpha - 0.04 - Added sprinting mechanics
Overhauled the ability to move and added a sprinting/resting mechanic that changed the game completely. This is now a challenging and "lively" game play that fits perfectly with the way Bubba's story lore is (An old assassin that prefers to end fights ASAP with the fewest, yet swiftest & precise movements as possible).