Beta - 1.25.5 - Boba pot code finally accepts boba bubble ingredients!
Fixed the logic that was stopping the boba pot from receiving ingredients! Turns out, it was because the save file didn't have the correct values and I didn't make a recent save with the maximum starting capacity of the pot to have 10 ingredients in it! Thus, the pot always thought it was "full" with the max ingredients of 0! (Also the tong check logic was slightly wrong and needed to check if the player was holding >=2 items).