Beta - 1.31.1 - MixerLogic&AbilitytoMix
Added some mixer logic to allow/deny the player to mix specific drink combinations. For example. if a customer wants a Pandan boba tea with boba, milk, and a green tea base, the player can craft a drink with such and use the mixer to blend them together! On another case, if the customer simply wanted water, the mixer allows you to make a water cup when you hit the "Mix" Button. However, I still need to apply the logic for what drink comes out of the mixer and to understand what drinks are being made!