Beta - 1.40.4 - Mat script - Drink order
Mat script finally orders the drinks in a FIFO array. Next step is to generate proper visual order of the drinks on the sell mat.

Minor fix on the UUID generation for single-item drinks causing an error and not generating a drink with one ingredient (I.E Oolong tea or Green Tea only drinks).