Beta - 1.40.45 - Mat script - Visual drink order
Made a way for the drinks to be shown in their LIFO stack from left->right where the right most drink is the oldest drink placed in the mat to be sold. This echoes real world perspective of customers (from a reverse POV of the barista's counter) were they usually grab the left most drink in a line of drinks in a coffee/boba shop, with the stigma is that drink is theirs's if they're called up to pick up their orders.