Beta - 1.86.1 Customer Order Pickups / Back to the grind
Added some animations and interactions
for when the customer comes back to "pick up" Their drinks. You need to still updated the "TAB" or "ORDERTAB" UI to vanish / update the current customers still waiting for their drinks. Also, you need find a way for the customers to have built in "cool downs" to prevent the animations from stacking.